Separate admin and island commands by package.
Easier to maintain.
