+++ src/main/java/world/bentobox/aoneblock/commands/island/IslandAboutCommand.java
@@ -1,4 +1,4 @@
-package world.bentobox.aoneblock.commands;
+package world.bentobox.aoneblock.commands.island;
 
 import java.util.List;
