+++ src/main/java/world/bentobox/aoneblock/commands/island/IslandCountCommand.java
@@ -1,4 +1,4 @@
-package world.bentobox.aoneblock.commands;
+package world.bentobox.aoneblock.commands.island;
 
 import java.util.List;
 import java.util.Objects;
